+++ src/main.py
@@ -4,10 +4,9 @@ Made by IMightBeMe and IMayBeMe on replit.com.
 Source code is available on github at https://github.com/IMightBeMe/Scythe and replit at https://replit.com/@IMightBeMe/Scythe?v=1
 '''
 
-from scythelexer import SLexer
-from scytheparser import SParser
-from scytheexecute import SExecute
-import glob
+from scythe_lexer import SLexer
+from scythe_parser import SParser
+from scythe_execute import SExecute
 
 
 if __name__ == '__main__':
