+++ src/scythe_lexer.py
